[REFACTOR] 코드리뷰 피드백 적용, dataSource 삭제 및 합침 #20
