+++ POLZZAK/Common/TabFilter/ViewProtocol/FooterViewUpdatable.swift
@@ -1,0 +1,12 @@
+//
+//  FooterViewUpdatable.swift
+//  POLZZAK
+//
+//  Created by 이정환 on 2023/08/29.
+//
+
+import Foundation
+
+protocol FooterViewUpdatable: AnyObject {
+    func updateCurrentCount(with count: Int)
+}
