+++ POLZZAK/Main/Cell/StampBoardFooterView.swift
@@ -62,7 +62,9 @@ extension StampBoardFooterView {
     func configure(with total: Int) {
         totalPage.text = "\(total)"
     }
-    
+}
+
+extension StampBoardFooterView: FooterViewUpdatable {
     func updateCurrentCount(with count: Int) {
         guard let totalCountText = totalPage.text else { return }
         guard let totalCount = Int(totalCountText) else { return }
@@ -72,4 +74,3 @@ extension StampBoardFooterView {
         }
     }
 }
-
